requirements.txt
asgiref==3.5.0
Django==4.0.4
djangorestframework==3.13.1
mysqlclient==2.0.1
django-mysql==3.9
django-cors-headers==3.5.0
pika==1.1.0